[Hanabi] Flag more cards as useless
For the time being, this does not show any
performance improvement: we lose as much
as we win:

    Computed scores in 1.988762
    Before: avg:27.717600, min:19, max:30
    After: avg:27.717600, min:19, max:30
    Losses -2 4 [-2, -1, -1, -1]
    Gains 2 4 [1, 1, 1, 2]

It was interesting anyway to see that
the counter were working pretty much
with pure luck and could not work with
newly instanciated objects. In order to
fix this, __eq__ and __hash__ need to be
implemented.